Fix missing cstdint include issues
ST_HAVE_INT64 is defined in st_config.h, but that's not being included directly in st_string.h. It gets pulled in by way of st_string_priv.h -> st_charbuffer.h, but the end result is that this initial ifdef was failing to include <cstdint> (because ST_HAVE_INT64 was not defined) but the ifdefs later in this header were passing (because at that point ST_HAVE_INT64 has been defined by later includes). This led to compilation errors about the use of uint64_t.