Fix for find (and related operations) reading past the end of the string.
Fixes #29.

This also fixes a unit test which previously expected the bad behavior.
Now, the `find_last()` functions behave as documented; they will only look
for a match where the whole substring is within the first `max` characters,
instead of returning matches which *started* in the first `max` characters
but extended into the rest of the string.