Re-vendor gtest (@ 1.13.0)
The submodule ended up causing a bunch of issues because it's not
included in zip/tarball downloads from GitHub and isn't cloned by
default, which led to build errors for things wanting to consume
string-theory as a dependency (since the CMake project here expects
gtest to exist and doesn't gracefully handle the case where it does
not).

Since we're essentially frozen at an old tag of gtest anyways, it
doesn't seem that bad to just go back to vendoring it like we used to.